refactor: DispatcherServlet 구조 변경 및 레거시 코드 제거
