--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/ManualHandlerAdapter.java
@@ -1,20 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.view.JspView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class ManualHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean accept(Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(Object handler, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        return ((Controller) handler).execute(request, response);
-    }
-}
